# docs/swagger.yaml
basePath: /
definitions:
  api_go.Author:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  api_go.AuthorInput:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  api_go.AuthorResponse:
    properties:
      book:
        items:
          $ref: '#/definitions/api_go.Book'
        type: array
      id:
        type: integer
      name:
        type: string
    type: object
  api_go.Book:
    properties:
      author_id:
        type: integer
      id:
        type: integer
      title:
        type: string
      years:
        type: integer
    type: object
  api_go.BookInput:
    properties:
      author_id:
        type: integer
      title:
        type: string
      years:
        type: integer
    required:
    - author_id
    - title
    - years
    type: object
  api_go.BookResponse:
    properties:
      author:
        $ref: '#/definitions/api_go.Author'
      id:
        type: integer
      title:
        type: string
      years:
        type: integer
    type: object
  api_go.User:
    properties:
      _:
        type: integer
      email:
        type: string
      is_active:
        type: boolean
      password:
        type: string
    required:
    - email
    - password
    type: object
  handler.GetBooksResponse:
    properties:
      result:
        items:
          $ref: '#/definitions/api_go.BookResponse'
        type: array
    type: object
  handler.GetUsersResponse:
    properties:
      result:
        items:
          $ref: '#/definitions/api_go.User'
        type: array
    type: object
  handler.errorResponse:
    properties:
      message:
        type: string
    type: object
  handler.signInInput:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
host: localhost:8000
info:
  contact: {}
  description: API
  title: App API
  version: "1.1"
paths:
  /admin/api/author:
    get:
      consumes:
      - application/json
      description: get authors
      operationId: get-authors
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: get authors
      tags:
      - author
    post:
      consumes:
      - application/json
      description: create author
      operationId: create-author
      parameters:
      - description: author info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/api_go.AuthorInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: create author
      tags:
      - author
  /admin/api/author/{id}:
    get:
      consumes:
      - application/json
      description: get author
      operationId: get-author
      parameters:
      - description: Author ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_go.AuthorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: get author by id
      tags:
      - author
  /admin/api/book:
    get:
      consumes:
      - application/json
      description: get books
      operationId: get-books
      parameters:
      - description: Author id
        in: query
        name: authorId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.GetBooksResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: get books
      tags:
      - book
    post:
      consumes:
      - application/json
      description: create book
      operationId: create-book
      parameters:
      - description: book info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/api_go.BookInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: create book
      tags:
      - book
  /admin/api/book/{id}:
    get:
      consumes:
      - application/json
      description: get book
      operationId: get-book
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_go.BookResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: get book by id
      tags:
      - book
  /admin/api/user:
    get:
      consumes:
      - application/json
      description: get all user
      operationId: get-all-user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.GetUsersResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: get users
      tags:
      - user
  /admin/api/user/{id}:
    get:
      consumes:
      - application/json
      description: get user
      operationId: get-user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_go.User'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: get user by id
      tags:
      - user
  /auth/sign-in:
    post:
      consumes:
      - application/json
      description: login
      operationId: login
      parameters:
      - description: credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.signInInput'
      produces:
      - application/json
      responses:
        "200":
          description: token
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: SignIn
      tags:
      - auth
  /auth/sign-up:
    post:
      consumes:
      - application/json
      description: registration
      operationId: registration
      parameters:
      - description: user info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/api_go.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: SignUp
      tags:
      - auth
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"